修复Tag开启列表读取的类型错误 / 空指针测试

+++ src/io/kurumi/ntt/spam/SpamTag.java
@@ -47,7 +47,7 @@ public class SpamTag extends AIIdDataModel {
 
         for (int index = 0; index < enableArray.size(); index++) {
 
-            enableArray.add(enableArray.getInt(index));
+            enableArray.add(enableArray.getLong(index));
 
         }
